Reapply "Modified and Added: File changes all across the repo and organized the whole folder structure"
This reverts commit 4cb1d4c9167a757071387d55f6793e181a02ad83.
