+++ about.html
@@ -1,0 +1,90 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>About Us | Ruler Tours</title>
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css"
+    />
+    <link rel="stylesheet" href="/css/global.css" />
+    <link rel="stylesheet" href="/css/about.css" />
+    <link rel="stylesheet" href="/css/style.css" />
+    <link rel="stylesheet" href="/css/footer.css" />
+    <script src="/javascript/auth-guard.js"></script>
+  </head>
+
+  <body>
+    <main class="about-container">
+      <section id="our-mission">
+        <h1 style="border-left: 5px solid gold; padding-left: 15px">
+          Our Mission
+        </h1>
+        <p>
+          At Ruler Tours, our mission is to redefine travel in North Bengal...
+        </p>
+      </section>
+
+      <div class="mission-vision-grid">
+        <div class="mission-box">
+          <h2>Our Vision</h2>
+          <p>
+            We envision a future where every traveler can "Rule their World"...
+          </p>
+        </div>
+      </div>
+    </main>
+
+    <footer class="footer">
+      <div class="footer-container">
+        <div class="footer-brand">
+          <h2>Ruler Tours</h2>
+          <p>
+            Premium tours and vehicle rentals crafted for modern explorers.
+            Travel with confidence and class.
+          </p>
+          <div class="footer-social">
+            <a href="#"><i class="fab fa-facebook"></i></a>
+            <a href="#"><i class="fab fa-instagram"></i></a>
+            <a href="#"><i class="fab fa-twitter"></i></a>
+          </div>
+        </div>
+
+        <div class="footer-links">
+          <h3>Quick Links</h3>
+          <ul>
+            <li><a href="/index.html">Home</a></li>
+            <li><a href="/destinations.html">Tours</a></li>
+            <li><a href="/rentals.html">Rentals</a></li>
+            <li><a href="/about.html">About</a></li>
+            <li><a href="/contact.html">Contact</a></li>
+          </ul>
+        </div>
+
+        <div class="footer-links">
+          <h3>Services</h3>
+          <ul>
+            <li><a href="/destinations.html">Luxury Tours</a></li>
+            <li><a href="/rentals.html">Car Rentals</a></li>
+            <li><a href="/contact.html">Corporate Trips</a></li>
+            <li><a href="/package.html">Custom Packages</a></li>
+          </ul>
+        </div>
+
+        <div class="footer-contact">
+          <h3>Contact Us</h3>
+          <p><i class="fas fa-envelope"></i> info@rulertours.com</p>
+          <p><i class="fas fa-phone"></i> +91 Your Number</p>
+          <p><i class="fas fa-map-marker-alt"></i> City, India</p>
+        </div>
+      </div>
+
+      <div class="footer-bottom">
+        <p>© 2026 Ruler Tours. All rights reserved.</p>
+      </div>
+      <script src="/javascript/main.js"></script>
+      <script src="/javascript/script.js"></script>
+    </footer>
+  </body>
+</html>
