+++ css/global.css
@@ -2,34 +2,34 @@
 @import url("footer.css");
 
 :root {
-    --primary: #333;
-    --accent: #007BFF;
-    --light: #f4f4f4;
-    --white: #ffffff;
+  --primary: #333;
+  --accent: #007bff;
+  --light: #f4f4f4;
+  --white: #ffffff;
 }
 
 body {
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    margin: 0;
-    line-height: 1.6;
-    color: var(--primary);
+  font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
+  margin: 0;
+  line-height: 1.6;
+  color: var(--primary);
 }
 
 /* NAVIGATION BAR */
 header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 1rem 5%;
-    border-bottom: 2px solid var(--light);
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 1rem 5%;
+  border-bottom: 2px solid var(--light);
 }
 
 /* Logo Styling */
 .site-logo {
-    height: 50px;
-    /* Adjust this to fit your design */
-    width: auto;
-    display: block;
+  height: 50px;
+  /* Adjust this to fit your design */
+  width: auto;
+  display: block;
 }
 
 /* HOOK: Dark Mode Placeholder 
@@ -37,163 +37,163 @@ header {
    and swap the logo to light_logo.png
 */
 body.dark-theme {
-    --bg-color: #1a1a1a;
-    --primary-color: #ffffff;
+  --bg-color: #1a1a1a;
+  --primary-color: #ffffff;
 }
 
 .logo {
-    font-size: 1.5rem;
-    font-weight: bold;
-    text-decoration: none;
-    color: var(--primary);
+  font-size: 1.5rem;
+  font-weight: bold;
+  text-decoration: none;
+  color: var(--primary);
 }
 
 nav ul {
-    display: flex;
-    list-style: none;
-    gap: 20px;
+  display: flex;
+  list-style: none;
+  gap: 20px;
 }
 
 nav a {
-    text-decoration: none;
-    color: var(--primary);
-    font-weight: 500;
+  text-decoration: none;
+  color: var(--primary);
+  font-weight: 500;
 }
 
 /* COMMON LAYOUT */
 section {
-    padding: 2rem 5%;
+  padding: 2rem 5%;
 }
 
 .grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-    gap: 20px;
-    margin-top: 20px;
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+  gap: 20px;
+  margin-top: 20px;
 }
 
 .card {
-    border: 1px solid #ddd;
-    padding: 15px;
-    border-radius: 8px;
-    text-align: center;
+  border: 1px solid #ddd;
+  padding: 15px;
+  border-radius: 8px;
+  text-align: center;
 }
 
 footer {
-    text-align: center;
-    padding: 2rem;
-    background: var(--light);
-    margin-top: 50px;
+  text-align: center;
+  padding: 2rem;
+  background: var(--light);
+  margin-top: 50px;
 }
 
-/* NAVBAR & PROFILE (From style.css) */
+/* NAVBAR & PROFILE (From /css/style.css) */
 .navbar {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 15px 60px;
-    position: sticky;
-    top: 0;
-    z-index: 1000;
-    background: linear-gradient(to right, #2E7D32, #FFB74D);
-    color: white;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 15px 60px;
+  position: sticky;
+  top: 0;
+  z-index: 1000;
+  background: linear-gradient(to right, #2e7d32, #ffb74d);
+  color: white;
 }
 
 .nav-left,
 .logo-link {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    text-decoration: none;
-    color: inherit;
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  text-decoration: none;
+  color: inherit;
 }
 
 .logo {
-    height: 42px;
+  height: 42px;
 }
 
 .nav-links {
-    display: flex;
-    gap: 35px;
-    list-style: none;
-    position: absolute;
-    left: 50%;
-    transform: translateX(-50%);
+  display: flex;
+  gap: 35px;
+  list-style: none;
+  position: absolute;
+  left: 50%;
+  transform: translateX(-50%);
 }
 
 .nav-links a {
-    color: white;
-    text-decoration: none;
-    font-weight: 500;
-    transition: 0.3s;
+  color: white;
+  text-decoration: none;
+  font-weight: 500;
+  transition: 0.3s;
 }
 
 .nav-right {
-    display: flex;
-    align-items: center;
-    gap: 15px;
+  display: flex;
+  align-items: center;
+  gap: 15px;
 }
 
 #theme-toggle {
-    padding: 6px 14px;
-    border: none;
-    border-radius: 20px;
-    cursor: pointer;
-    background: white;
+  padding: 6px 14px;
+  border: none;
+  border-radius: 20px;
+  cursor: pointer;
+  background: white;
 }
 
 .hamburger {
-    display: none;
-    font-size: 26px;
-    cursor: pointer;
+  display: none;
+  font-size: 26px;
+  cursor: pointer;
 }
 
 .profile-container {
-    position: relative;
-    cursor: pointer;
-    display: none;
+  position: relative;
+  cursor: pointer;
+  display: none;
 }
 
 .profile-avatar {
-    width: 40px;
-    height: 40px;
-    border-radius: 50%;
-    border: 2px solid white;
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  border: 2px solid white;
 }
 
 .profile-dropdown {
-    position: absolute;
-    top: 55px;
-    right: 0;
-    background: white;
-    min-width: 180px;
-    border-radius: 12px;
-    box-shadow: 0 10px 25px rgba(0, 0, 0, 0.1);
-    display: none;
-    flex-direction: column;
-    padding: 10px 0;
-    z-index: 1100;
+  position: absolute;
+  top: 55px;
+  right: 0;
+  background: white;
+  min-width: 180px;
+  border-radius: 12px;
+  box-shadow: 0 10px 25px rgba(0, 0, 0, 0.1);
+  display: none;
+  flex-direction: column;
+  padding: 10px 0;
+  z-index: 1100;
 }
 
 .profile-dropdown.active {
-    display: flex;
+  display: flex;
 }
 
 .profile-dropdown a {
-    padding: 12px 20px;
-    text-decoration: none;
-    color: #333;
-    display: flex;
-    align-items: center;
-    gap: 10px;
+  padding: 12px 20px;
+  text-decoration: none;
+  color: #333;
+  display: flex;
+  align-items: center;
+  gap: 10px;
 }
 
 @media (max-width: 768px) {
-    .hamburger {
-        display: block;
-    }
+  .hamburger {
+    display: block;
+  }
 
-    .nav-links {
-        display: none;
-    }
+  .nav-links {
+    display: none;
+  }
 }
