+++ destinations.html
@@ -1,0 +1,77 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <title>Destinations | Ruler Tours</title>
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css"
+    />
+    <link rel="stylesheet" href="../css/global.css" />
+    <link rel="stylesheet" href="../css/destinations.css" />
+    <link rel="stylesheet" href="../css/footer.css" />
+    <script src="../javascript/auth-guard.js"></script>
+  </head>
+
+  <body>
+    <section>
+      <h1 style="text-align: center; font-family: serif">
+        North Bengal Destinations
+      </h1>
+
+      <div id="destinations-grid"></div>
+    </section>
+
+    <footer class="footer">
+      <div class="footer-container">
+        <div class="footer-brand">
+          <h2>Ruler Tours</h2>
+          <p>
+            Premium tours and vehicle rentals crafted for modern explorers.
+            Travel with confidence and class.
+          </p>
+          <div class="footer-social">
+            <a href="#"><i class="fab fa-facebook"></i></a>
+            <a href="#"><i class="fab fa-instagram"></i></a>
+            <a href="#"><i class="fab fa-twitter"></i></a>
+          </div>
+        </div>
+
+        <div class="footer-links">
+          <h3>Quick Links</h3>
+          <ul>
+            <li><a href="../index.html">Home</a></li>
+            <li><a href="/destinations.html">Tours</a></li>
+            <li><a href="/rentals.html">Rentals</a></li>
+            <li><a href="/about.html">About</a></li>
+            <li><a href="/contact.html">Contact</a></li>
+          </ul>
+        </div>
+
+        <div class="footer-links">
+          <h3>Services</h3>
+          <ul>
+            <li><a href="/destinations.html">Luxury Tours</a></li>
+            <li><a href="/rentals.html">Car Rentals</a></li>
+            <li><a href="/contact.html">Corporate Trips</a></li>
+            <li><a href="/package.html">Custom Packages</a></li>
+          </ul>
+        </div>
+
+        <div class="footer-contact">
+          <h3>Contact Us</h3>
+          <p><i class="fas fa-envelope"></i> info@rulertours.com</p>
+          <p><i class="fas fa-phone"></i> +91 Your Number</p>
+          <p><i class="fas fa-map-marker-alt"></i> City, India</p>
+        </div>
+      </div>
+
+      <div class="footer-bottom">
+        <p>© 2026 Ruler Tours. All rights reserved.</p>
+      </div>
+    </footer>
+    <script src="/javascript/tour-data.js"></script>
+    <script src="/javascript/main.js"></script>
+    <script src="/javascript/script.js"></script>
+  </body>
+</html>
