+++ footer.html
@@ -14,21 +14,21 @@
     <div class="footer-links">
       <h3>Quick Links</h3>
       <ul>
-        <li><a href="../index.html">Home</a></li>
-        <li><a href="./destinations.html">Tours</a></li>
-        <li><a href="./rentals.html">Rentals</a></li>
-        <li><a href="./about.html">About</a></li>
-        <li><a href="./contact.html">Contact</a></li>
+        <li><a href="/index.html">Home</a></li>
+        <li><a href="/destinations.html">Tours</a></li>
+        <li><a href="/rentals.html">Rentals</a></li>
+        <li><a href="/about.html">About</a></li>
+        <li><a href="/contact.html">Contact</a></li>
       </ul>
     </div>
 
     <div class="footer-links">
       <h3>Services</h3>
       <ul>
-        <li><a href="./destinations.html">Luxury Tours</a></li>
-        <li><a href="./rentals.html">Car Rentals</a></li>
-        <li><a href="./contact.html">Corporate Trips</a></li>
-        <li><a href="./package.html">Custom Packages</a></li>
+        <li><a href="/destinations.html">Luxury Tours</a></li>
+        <li><a href="/rentals.html">Car Rentals</a></li>
+        <li><a href="/contact.html">Corporate Trips</a></li>
+        <li><a href="/package.html">Custom Packages</a></li>
       </ul>
     </div>
