--- forgot-password.html
@@ -1,56 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Forgot Password - Ruler Tours</title>
-    <link rel="icon" type="image/png" href="assets/light_logo.png">
-    <link rel="stylesheet" href="style.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
-    <link
-        href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600&family=Outfit:wght@500;600;700&display=swap"
-        rel="stylesheet">
-    <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/npm/toastify-js/src/toastify.min.css">
-    <script src="javascript/auth-guard.js"></script>
-</head>
-
-<body>
-    <section class="auth-section">
-        <div class="auth-container">
-            <h2>Forgot Password</h2>
-            <p style="margin-bottom: 20px; color: var(--text-muted);">Enter your email to receive a 6-digit OTP.</p>
-            <form id="forgotPasswordForm">
-                <div class="form-group">
-                    <input type="email" id="forgotPassEmail" placeholder="Email Address" required>
-                </div>
-                <button type="submit" class="btn-primary" style="width: 100%;">Send OTP</button>
-            </form>
-            <p><a href="login.html">Back to Login</a></p>
-        </div>
-    </section>
-
-    <footer class="footer">
-        <div class="footer-content">
-            <div>
-                <h3>Ruler Tours</h3>
-                <p>Discover rural beauty & eco adventures</p>
-            </div>
-            <div>
-                <h4>Quick Links</h4>
-                <p><a href="tours.html">Tours</a></p>
-                <p><a href="rentals.html">Rentals</a></p>
-            </div>
-            <div>
-                <h4>Contact</h4>
-                <p>+91 98765 43210</p>
-                <p>info@rulertours.com</p>
-            </div>
-        </div>
-    </footer>
-
-    <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/toastify-js"></script>
-    <script src="script.js"></script>
-</body>
-
-</html>
