--- html/contact.html
@@ -1,79 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <title>Contact Us - Ruler Tours</title>
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
-  <link rel="stylesheet" href="../css/global.css">
-  <link rel="stylesheet" href="../style.css">
-  <link rel="stylesheet" href="../css/footer.css">
-  <script src="../javascript/auth-guard.js"></script>
-</head>
-
-<body>
-  <section>
-    <h2>Get in Touch</h2>
-    <form method="POST"
-      style="max-width: 400px; display: flex; flex-direction: column; gap: 10px;" id="contactForm">
-
-      <input name="name" type="text" placeholder="Your Name" style="padding: 10px;">
-      <input name="email" type="email" placeholder="Your Email" style="padding: 10px;">
-      <textarea name="message" placeholder="Your Message" rows="5" style="padding: 10px;"></textarea>
-      <button type="submit" style="padding: 10px; background: black; color: white; cursor: pointer;">Send
-        Message</button>
-    </form>
-  </section>
-
-  <footer class="footer">
-    <div class="footer-container">
-
-      <div class="footer-brand">
-        <h2>Ruler Tours</h2>
-        <p>Premium tours and vehicle rentals crafted for modern explorers. Travel with confidence and class.</p>
-        <div class="footer-social">
-          <a href="#"><i class="fab fa-facebook"></i></a>
-          <a href="#"><i class="fab fa-instagram"></i></a>
-          <a href="#"><i class="fab fa-twitter"></i></a>
-        </div>
-      </div>
-
-      <div class="footer-links">
-        <h3>Quick Links</h3>
-        <ul>
-          <li><a href="../index.html">Home</a></li>
-          <li><a href="html/destinations.html">Tours</a></li>
-          <li><a href="html/rentals.html">Rentals</a></li>
-          <li><a href="html/about.html">About</a></li>
-          <li><a href="html/contact.html">Contact</a></li>
-        </ul>
-      </div>
-
-      <div class="footer-links">
-        <h3>Services</h3>
-        <ul>
-          <li><a href="html/destinations.html">Luxury Tours</a></li>
-          <li><a href="html/rentals.html">Car Rentals</a></li>
-          <li><a href="html/contact.html">Corporate Trips</a></li>
-          <li><a href="html/package.html">Custom Packages</a></li>
-        </ul>
-      </div>
-
-      <div class="footer-contact">
-        <h3>Contact Us</h3>
-        <p><i class="fas fa-envelope"></i> info@rulertours.com</p>
-        <p><i class="fas fa-phone"></i> +91 Your Number</p>
-        <p><i class="fas fa-map-marker-alt"></i> City, India</p>
-      </div>
-
-    </div>
-
-    <div class="footer-bottom">
-      <p>© 2026 Ruler Tours. All rights reserved.</p>
-    </div>
-
-    <script src="../javascript/config-form.js"></script>
-    <script src="../script.js"></script>
-</body>
-
-</html>
