--- html/navbar.html
@@ -1,35 +1,0 @@
-<nav class="navbar" id="main-navbar">
-    <div class="nav-left">
-        <a href="index.html" class="logo-link">
-            <img src="assets/light_logo.png" id="logo" class="logo" alt="Ruler Tours Logo">
-            <h2>Ruler Tours</h2>
-        </a>
-    </div>
-
-    <ul class="nav-links" id="nav-links">
-        <li><a href="index.html" data-page="index.html">Home</a></li>
-        <li><a href="tours.html" data-page="tours.html">Tours</a></li>
-        <li><a href="rentals.html" data-page="rentals.html">Rentals</a></li>
-        <li><a href="html/about.html" data-page="about.html">About</a></li>
-        <li><a href="html/contact.html" data-page="contact.html">Contact</a></li>
-    </ul>
-
-    <div class="nav-right">
-        <button id="theme-toggle">🌙</button>
-
-        <!-- Auth Buttons -->
-        <a href="login.html" id="auth-btn" class="btn-login">Login</a>
-        <a href="signup.html" id="signup-btn" class="btn-primary" style="padding: 8px 20px; font-size: 0.9rem;">Sign
-            Up</a>
-
-        <!-- Profile (Hidden by default) -->
-        <div class="profile-container" id="profile-trigger">
-            <img src="" class="profile-avatar" alt="User">
-            <div class="profile-dropdown" id="profile-dropdown">
-                <a href="#" id="logout-btn"><i class="fas fa-sign-out-alt"></i> Logout</a>
-            </div>
-        </div>
-
-        <div class="hamburger" id="hamburger">☰</div>
-    </div>
-</nav>
