--- html/rentals.html
@@ -1,70 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <title>Rentals - Ruler Tours</title>
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
-  <link rel="stylesheet" href="../css/global.css">
-  <link rel="stylesheet" href="../style.css">
-  <link rel="stylesheet" href="../css/footer.css">
-  <script src="../javascript/auth-guard.js"></script>
-</head>
-
-<body>
-  <section>
-    <h2>Rent a Vehicle</h2>
-    <div class="grid">
-      <div class="card">
-        <h3>🚗 SUV - Toyota Land Cruiser</h3>
-        <p>$120 / Day</p>
-        <button>Rent Now</button>
-      </div>
-      <div class="card">
-        <h3>🏍️ Adventure Bike - BMW GS</h3>
-        <p>$60 / Day</p>
-        <button>Rent Now</button>
-      </div>
-    </div>
-    </div>
-
-    <div class="footer-links">
-      <h3>Quick Links</h3>
-      <ul>
-        <li><a href="../index.html">Home</a></li>
-        <li><a href="html/destinations.html">Tours</a></li>
-        <li><a href="html/rentals.html">Rentals</a></li>
-        <li><a href="html/about.html">About</a></li>
-        <li><a href="html/contact.html">Contact</a></li>
-      </ul>
-    </div>
-
-    <div class="footer-links">
-      <h3>Services</h3>
-      <ul>
-        <li><a href="html/destinations.html">Luxury Tours</a></li>
-        <li><a href="html/rentals.html">Car Rentals</a></li>
-        <li><a href="html/contact.html">Corporate Trips</a></li>
-        <li><a href="html/package.html">Custom Packages</a></li>
-      </ul>
-    </div>
-
-    <div class="footer-contact">
-      <h3>Contact Us</h3>
-      <p><i class="fas fa-envelope"></i> info@rulertours.com</p>
-      <p><i class="fas fa-phone"></i> +91 Your Number</p>
-      <p><i class="fas fa-map-marker-alt"></i> City, India</p>
-    </div>
-
-    </div>
-
-    </div>
-
-    <div class="footer-bottom">
-      <p>© 2026 Ruler Tours. All rights reserved.</p>
-    </div>
-    <script src="../script.js"></script>
-    </footer>
-</body>
-
-</html>
