+++ index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Ruler Tours</title>
     <link rel="icon" type="image/png" href="assets/light_logo.png" />
-    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="/css/style.css" />
     <link
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css"
@@ -120,12 +120,15 @@
         </div>
 
         <div class="contact-form-box">
-          <form
-            method="POST"
-            id="contactForm"
-          >
+          <form method="POST" id="contactForm">
             <div class="form-group">
-              <input name="name" type="text" id="name" placeholder="Your Name" required />
+              <input
+                name="name"
+                type="text"
+                id="name"
+                placeholder="Your Name"
+                required
+              />
             </div>
 
             <div class="form-group">
@@ -160,7 +163,6 @@
 
             <button type="submit" class="btn-primary">Send Message</button>
           </form>
-
         </div>
       </div>
     </section>
@@ -193,7 +195,7 @@
       <p class="copyright">&copy; 2026 Ruler Tours. All rights reserved.</p>
     </footer>
 
-    <script src="./javascript/config-form.js"></script>
-    <script src="script.js"></script>
+    <script src="/javascript/config-form.js"></script>
+    <script src="/javascript/script.js"></script>
   </body>
 </html>
