+++ javascript/rental-logic.js
@@ -1,7 +1,7 @@
 // Logic specific to the Vehicle Rental page
 console.log("Rental Logic Loaded");
 
-import {vehicleData} from "../rental-vehicle-data.js"
+import {vehicleData} from "/db/rental-vehicle-data.js"
 
 // console.log(vehicleData);
