+++ javascript/script.js
@@ -15,7 +15,7 @@ async function loadNavbarComponent() {
         <nav class="navbar" id="main-navbar">
             <div class="nav-left">
                 <a href="index.html" class="logo-link">
-                    <img src="assets/light_logo.png" id="logo" class="logo" alt="Ruler Tours Logo">
+                    <img src="/assets/light_logo.png" id="logo" class="logo" alt="Ruler Tours Logo">
                     <h2>Ruler Tours</h2>
                 </a>
             </div>
@@ -24,8 +24,8 @@ async function loadNavbarComponent() {
                 <li><a href="index.html" data-page="index.html">Home</a></li>
                 <li><a href="tours.html" data-page="tours.html">Tours</a></li>
                 <li><a href="rentals.html" data-page="rentals.html">Rentals</a></li>
-                <li><a href="html/about.html" data-page="about.html">About</a></li>
-                <li><a href="html/contact.html" data-page="contact.html">Contact</a></li>
+                <li><a href="about.html" data-page="about.html">About</a></li>
+                <li><a href="contact.html" data-page="contact.html">Contact</a></li>
             </ul>
 
             <div class="nav-right">
@@ -50,35 +50,6 @@ async function loadNavbarComponent() {
 
         // Insert navbar at top of body
         document.body.insertAdjacentHTML('afterbegin', navbarHTML);
-
-        // --- DYNAMIC LINK FIXING ---
-        const prefix = isInHtmlFolder ? '../' : '';
-        const navbar = document.querySelector('.navbar');
-        if (navbar) {
-            // Fix all <a> links in navbar
-            const links = navbar.querySelectorAll('a');
-            links.forEach(link => {
-                const href = link.getAttribute('href');
-                if (href && !href.startsWith('http') && !href.startsWith('#')) {
-                    const isTargetInHtmlFolder = href.startsWith('html/');
-                    const cleanHref = isTargetInHtmlFolder ? href.replace('html/', '') : href;
-
-                    if (isInHtmlFolder) {
-                        if (href.startsWith('html/')) {
-                            newHref = href.replace('html/', '');
-                        } else {
-                            newHref = '../' + href;
-                        }
-                    }
-                }
-            });
-
-            // Fix Logo Path
-            const logo = navbar.querySelector('#logo');
-            if (logo) {
-                logo.src = prefix + 'assets/light_logo.png';
-            }
-        }
 
         console.log('✅ Navbar injected locally and links adjusted');
         return true;
