+++ rentals.html
@@ -1,82 +1,75 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Ruler Tours</title>
-    <link rel="icon" type="image/png" href="assets/light_logo.png">
-    <link rel="stylesheet" href="style.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
+    <link rel="icon" type="image/png" href="assets/light_logo.png" />
+    <link rel="stylesheet" href="/css/style.css" />
     <link
-        href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600&family=Outfit:wght@500;600;700&display=swap"
-        rel="stylesheet">
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css"
+    />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600&family=Outfit:wght@500;600;700&display=swap"
+      rel="stylesheet"
+    />
     <script src="javascript/auth-guard.js"></script>
-</head>
+  </head>
 
-<body>
-    
+  <body>
     <section class="page-hero">
-        <h1>Premium Fleet Rentals</h1>
+      <h1>Premium Fleet Rentals</h1>
     </section>
 
     <div class="rental-cards">
-
-    <section class="section">
+      <section class="section">
         <h2>Luxury Cruisers</h2>
-    </section>
+      </section>
 
-    <section class="card-container" id="luxury-cruisers">
-
+      <section class="card-container" id="luxury-cruisers">
         <!-- rendered from js -->
+      </section>
 
-    </section>
-
-    <section class="section">
+      <section class="section">
         <h2>Bike Rentals</h2>
-    </section>
+      </section>
 
-    <section class="card-container" id="bike-rentals">
-
+      <section class="card-container" id="bike-rentals">
         <!-- rendered from js -->
-        
-    </section>
+      </section>
     </div>
-    
+
     <footer class="footer">
-        <div class="footer-content">
-
-            <div>
-                <h3>Ruler Tours</h3>
-                <p>Next-gen travel & premium itineraries</p>
-            </div>
-
-            <div>
-                <h4>Explore</h4>
-                <p><a href="tours.html">Tours</a></p>
-                <p><a href="rentals.html">Rentals</a></p>
-            </div>
-
-            <div>
-                <h4>Connect</h4>
-                <p>+1 800 555 RULER</p>
-                <p>info@rulertours.com</p>
-                <div class="social-links">
-                    <a href="#" target="_blank"><i class="fab fa-instagram"></i></a>
-                    <a href="#" target="_blank"><i class="fab fa-facebook-f"></i></a>
-                    <a href="#" target="_blank"><i class="fab fa-twitter"></i></a>
-                    <a href="#" target="_blank"><i class="fab fa-linkedin-in"></i></a>
-                </div>
-            </div>
-
+      <div class="footer-content">
+        <div>
+          <h3>Ruler Tours</h3>
+          <p>Next-gen travel & premium itineraries</p>
         </div>
 
-        <p class="copyright">&copy; 2026 Ruler Tours. All rights reserved.</p>
+        <div>
+          <h4>Explore</h4>
+          <p><a href="tours.html">Tours</a></p>
+          <p><a href="rentals.html">Rentals</a></p>
+        </div>
+
+        <div>
+          <h4>Connect</h4>
+          <p>+1 800 555 RULER</p>
+          <p>info@rulertours.com</p>
+          <div class="social-links">
+            <a href="#" target="_blank"><i class="fab fa-instagram"></i></a>
+            <a href="#" target="_blank"><i class="fab fa-facebook-f"></i></a>
+            <a href="#" target="_blank"><i class="fab fa-twitter"></i></a>
+            <a href="#" target="_blank"><i class="fab fa-linkedin-in"></i></a>
+          </div>
+        </div>
+      </div>
+
+      <p class="copyright">&copy; 2026 Ruler Tours. All rights reserved.</p>
     </footer>
 
-
-    <script type="module" src="./javascript/rental-logic.js"></script>
-    <script src="script.js"></script>
-</body>
-
+    <script type="module" src="/javascript/rental-logic.js"></script>
+    <script src="/javascript/script.js"></script>
+  </body>
 </html>
