--- reset-password.html
@@ -1,64 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Reset Password - Ruler Tours</title>
-    <link rel="icon" type="image/png" href="assets/light_logo.png">
-    <link rel="stylesheet" href="style.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
-    <link
-        href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600&family=Outfit:wght@500;600;700&display=swap"
-        rel="stylesheet">
-    <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/npm/toastify-js/src/toastify.min.css">
-    <script src="javascript/auth-guard.js"></script>
-</head>
-
-<body>
-    <section class="auth-section">
-        <div class="auth-container">
-            <h2>Reset Password</h2>
-            <p style="margin-bottom: 20px; color: var(--text-muted);">Enter the OTP sent to your email and your new
-                password.</p>
-            <form id="resetPasswordForm">
-                <div class="form-group">
-                    <input type="text" id="resetOtp" placeholder="6-digit OTP" required maxlength="6">
-                </div>
-                <div class="form-group">
-                    <input type="password" id="resetNewPassword" placeholder="New Password" required minlength="6">
-                </div>
-                <div class="form-group">
-                    <input type="password" id="resetConfirmPassword" placeholder="Confirm New Password" required
-                        minlength="6">
-                </div>
-                <button type="submit" class="btn-primary" style="width: 100%;">Reset Password</button>
-            </form>
-            <p><a href="forgot-password.html">Resend OTP?</a></p>
-        </div>
-    </section>
-
-    <footer class="footer">
-        <div class="footer-content">
-            <div>
-                <h3>Ruler Tours</h3>
-                <p>Discover rural beauty & eco adventures</p>
-            </div>
-            <div>
-                <h4>Quick Links</h4>
-                <p><a href="tours.html">Tours</a></p>
-                <p><a href="rentals.html">Rentals</a></p>
-            </div>
-            <div>
-                <h4>Contact</h4>
-                <p>+91 98765 43210</p>
-                <p>info@rulertours.com</p>
-            </div>
-        </div>
-    </footer>
-
-    <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/toastify-js"></script>
-    <script src="script.js"></script>
-</body>
-
-</html>
