--- reset-success.html
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Reset Successful - Ruler Tours</title>
-    <link rel="icon" type="image/png" href="assets/light_logo.png">
-    <link rel="stylesheet" href="style.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
-    <link
-        href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600&family=Outfit:wght@500;600;700&display=swap"
-        rel="stylesheet">
-    <script src="javascript/auth-guard.js"></script>
-</head>
-
-<body>
-    <section class="auth-section">
-        <div class="auth-container" style="text-align: center;">
-            <div style="font-size: 50px; color: #2E7D32; margin-bottom: 20px;">
-                <i class="fas fa-check-circle"></i>
-            </div>
-            <h2>Password Reset Successful</h2>
-            <p style="margin-bottom: 30px; color: var(--text-muted);">Your password has been reset successfully. You can
-                now login with your new password.</p>
-            <a href="login.html" class="btn-primary" style="display: block; width: 100%; text-decoration: none;">Login
-                Now</a>
-        </div>
-    </section>
-
-    <footer class="footer">
-        <div class="footer-content">
-            <div>
-                <h3>Ruler Tours</h3>
-                <p>Discover rural beauty & eco adventures</p>
-            </div>
-            <div>
-                <h4>Quick Links</h4>
-                <p><a href="tours.html">Tours</a></p>
-                <p><a href="rentals.html">Rentals</a></p>
-            </div>
-            <div>
-                <h4>Contact</h4>
-                <p>+91 98765 43210</p>
-                <p>info@rulertours.com</p>
-            </div>
-        </div>
-    </footer>
-
-    <script src="script.js"></script>
-</body>
-
-</html>
